remove `src/app/gui/streetview/streetviewservice.js`
- move `src/app/gui/streetview/streetviewservice.js` within `src/app/g3w-ol/controls/streetviewcontrol.js`
- remove unusued `StreetViewService` imports
