--- src/app/gui/streetview/streetviewservice.js
@@ -1,7 +1,0 @@
-/**
- * DEPRECATED: this file will be removed after v3.4 (use "services/streetview.js" instead)
- */
-
-const StreetViewService = require('services/streetview');
-
-module.exports = StreetViewService;
